Move updatable dashboard types to /models
