+++ src/models/dashboard/index.ts
@@ -18,7 +18,10 @@ export * from './is-dashboard';
 export * from './is-dashboard-search';
 export * from './raw-creatable-dashboard';
 export * from './raw-dashboard';
+export * from './raw-updatable-dashboard';
 export * from './to-dashboard';
 export * from './to-raw-creatable-dashboard';
 export * from './to-raw-creatable-dashboard-search';
 export * from './to-raw-creatable-dashboard-tile';
+export * from './to-raw-updatable-dashboard';
+export * from './updatable-dashboard';
