+++ src/models/dashboard/to-raw-updatable-dashboard.ts
@@ -1,0 +1,42 @@
+import { isUndefined } from 'lodash';
+import { omitUndefinedShallow } from '../../functions/utils';
+import { toRawNumericID } from '../../value-objects';
+import { toRawTimeframe } from '../timeframe';
+import { Dashboard } from './dashboard';
+import { RawUpdatableDashboard } from './raw-updatable-dashboard';
+import { toRawCreatableDashboardSearch } from './to-raw-creatable-dashboard-search';
+import { UpdatableDashboard } from './updatable-dashboard';
+
+export const toRawUpdatableDashboard = (updatable: UpdatableDashboard, current: Dashboard): RawUpdatableDashboard => ({
+	GIDs: (updatable.groupIDs ?? current.groupIDs).map(toRawNumericID),
+
+	Name: updatable.name ?? current.name,
+	Description: (isUndefined(updatable.description) ? current.description : updatable.description) ?? '',
+	Labels: updatable.labels ?? current.labels,
+
+	Data: {
+		timeframe: toRawTimeframe(updatable.timeframe ?? current.timeframe),
+
+		searches: (updatable.searches ?? current.searches).map(toRawCreatableDashboardSearch),
+
+		tiles: (updatable.tiles ?? current.tiles).map(t => ({
+			id: toRawNumericID(t.id),
+			title: t.title,
+			renderer: t.renderer,
+			span: { col: t.dimensions.columns, row: t.dimensions.rows, x: t.position.x, y: t.position.y },
+			searchesIndex: t.searchIndex,
+			rendererOptions: t.rendererOptions,
+		})),
+
+		liveUpdateInterval: (updatable.liveUpdate ?? current.liveUpdate).interval ?? undefined,
+		linkZooming: updatable.updateOnZoom ?? current.updateOnZoom,
+		grid: (() => {
+			const gridOptions = { ...current.gridOptions, ...(updatable.gridOptions ?? {}) };
+			return omitUndefinedShallow({
+				gutter: gridOptions.gutter ?? undefined,
+				margin: gridOptions.margin ?? undefined,
+			});
+		})(),
+		version: updatable.version?.major ?? current.version.major,
+	},
+});
